Remove signed/unsigned comparison warning.
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3-porting@11242 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_Image_Surface.cxx
@@ -240,7 +240,7 @@ static int find_slot() { // return an available slot to memorize an Fl_Image_Sur
     for (int num = 0; num < count; num++) {
         if (!offscreen_api_surface[num]) return num;
       }
-    if (count >= sizeof(offscreen_api_surface)/sizeof(Fl_Image_Surface*)) return -1;
+    if ((unsigned)count >= sizeof(offscreen_api_surface)/sizeof(Fl_Image_Surface*)) return -1;
     return count++;
   }
